fix error page tests
